requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
openai>=1.10.0
httpx>=0.26.0
jinja2>=3.1.3
loguru>=0.7.2
python-dotenv>=1.0.0
python-docx>=1.1.0
python-Levenshtein>=0.25.0
lxml>=5.0.0
